+++ UserVoiceSDK/src/com/uservoice/uservoicesdk/flow/SigninCallback.java
@@ -1,0 +1,6 @@
+package com.uservoice.uservoicesdk.flow;
+
+public abstract class SigninCallback {
+    public abstract void onSuccess();
+    public void onFailure() {}
+}