+++ UserVoiceSDK/src/com/uservoice/uservoicesdk/flow/SigninManager.java
@@ -2,10 +2,13 @@ package com.uservoice.uservoicesdk.flow;
 
 import android.support.v4.app.FragmentActivity;
 
+import android.widget.Toast;
+import com.uservoice.uservoicesdk.R;
 import com.uservoice.uservoicesdk.Session;
 import com.uservoice.uservoicesdk.babayaga.Babayaga;
 import com.uservoice.uservoicesdk.dialog.PasswordDialogFragment;
 import com.uservoice.uservoicesdk.dialog.SigninDialogFragment;
+import com.uservoice.uservoicesdk.flow.SigninCallback;
 import com.uservoice.uservoicesdk.model.AccessTokenResult;
 import com.uservoice.uservoicesdk.model.RequestToken;
 import com.uservoice.uservoicesdk.model.User;
@@ -13,23 +16,26 @@ import com.uservoice.uservoicesdk.rest.Callback;
 import com.uservoice.uservoicesdk.rest.RestResult;
 import com.uservoice.uservoicesdk.ui.DefaultCallback;
 
+import java.util.regex.Pattern;
+
 public class SigninManager {
 
-    private final Runnable callback;
+    private final SigninCallback callback;
     private String email;
     private String name;
     private final FragmentActivity activity;
     private boolean passwordOnly;
+    private Pattern emailFormat = Pattern.compile("\\A(\\w[-+.\\w!\\#\\$%&'\\*\\+\\-/=\\?\\^_`\\{\\|\\}~]*@([-\\w]*\\.)+[a-zA-Z]{2,9})\\z");
 
-    public static void signIn(FragmentActivity activity, Runnable callback) {
+    public static void signIn(FragmentActivity activity, SigninCallback callback) {
         new SigninManager(activity, null, null, callback).signIn();
     }
 
-    public static void signIn(FragmentActivity activity, String email, String name, Runnable callback) {
+    public static void signIn(FragmentActivity activity, String email, String name, SigninCallback callback) {
         new SigninManager(activity, email, name, callback).signIn();
     }
 
-    private SigninManager(FragmentActivity activity, String email, String name, Runnable callback) {
+    private SigninManager(FragmentActivity activity, String email, String name, SigninCallback callback) {
         this.activity = activity;
         this.email = email == null || email.trim().length() == 0 ? null : email;
         this.name = name == null || name.trim().length() == 0 ? null : name;
@@ -39,10 +45,13 @@ public class SigninManager {
     private void signIn() {
         User currentUser = Session.getInstance().getUser();
         if (currentUser != null && (email == null || email.equals(currentUser.getEmail()))) {
-            callback.run();
+            callback.onSuccess();
         } else if (Session.getInstance().getAccessToken() != null) {
             // If we have an access token but no user, they have signed in in this session. Don't prompt again.
-            callback.run();
+            callback.onSuccess();
+        } else if (email != null && !emailFormat.matcher(email).matches()) {
+            Toast.makeText(activity, R.string.uv_msg_bad_email_format, Toast.LENGTH_SHORT).show();
+            callback.onFailure();
         } else {
             email = email == null ? Session.getInstance().getEmail() : email;
             name = name == null ? Session.getInstance().getName() : name;
@@ -75,7 +84,7 @@ public class SigninManager {
                         Session.getInstance().setUser(model.getModel());
                         Session.getInstance().setAccessToken(activity, model.getAccessToken());
                         Babayaga.track(Babayaga.Event.IDENTIFY);
-                        callback.run();
+                        callback.onSuccess();
                     }
                 });
             }
@@ -96,7 +105,7 @@ public class SigninManager {
         this.passwordOnly = passwordOnly;
     }
 
-    public static void signinForSubscribe(FragmentActivity activity, String email, Runnable callback) {
+    public static void signinForSubscribe(FragmentActivity activity, String email, SigninCallback callback) {
         SigninManager manager = new SigninManager(activity, email, Session.getInstance().getName(), callback);
         manager.setPasswordOnly(true);
         manager.signIn();