+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add portuguese translation
 * fix a bug where UserVoice.init() would show an error message to the user if the server was unreachable
 * Prevent duplicat ideas or tickets if the user taps submit multiple times
+* Add email validation
 
 ## 1.0.0